Add Unifier application and move config to core... and indexing... and queue.... and ingestion

+++ src/harvest/Fip.Strive.Harvester.Application/Features/Assimilate/Services/AssimilateClassifyWorker.cs
@@ -1,13 +1,12 @@
-using Fip.Strive.Core.Domain.Schemas.Queue.Enums;
-using Fip.Strive.Core.Domain.Schemas.Queue.Models;
-using Fip.Strive.Core.Domain.Schemas.Queue.Models.Signals;
-using Fip.Strive.Harvester.Application.Core.Queue.Components;
 using Fip.Strive.Harvester.Application.Features.Assimilate.Services.Contracts;
+using Fip.Strive.Harvester.Domain.Signals;
+using Fip.Strive.Queue.Application.Components;
+using Fip.Strive.Queue.Domain.Models;
 
 namespace Fip.Strive.Harvester.Application.Features.Assimilate.Services;
 
 public class AssimilateClassifyWorker(IAssimilationService assimilator)
-    : QueueWorker(SignalTypes.TypedSignal)
+    : QueueWorker((int)SignalTypes.TypedSignal)
 {
     public override async Task ProcessAsync(JobDetails job, CancellationToken ct)
     {
