+++ src/harvest/Fip.Strive.Harvester.Application/Features/Classify/Services/ClassifyExpandWorker.cs
@@ -1,12 +1,11 @@
-using Fip.Strive.Core.Domain.Schemas.Queue.Enums;
-using Fip.Strive.Core.Domain.Schemas.Queue.Models;
-using Fip.Strive.Core.Domain.Schemas.Queue.Models.Signals;
-using Fip.Strive.Harvester.Application.Core.Queue.Components;
 using Fip.Strive.Harvester.Application.Features.Classify.Services.Contracts;
+using Fip.Strive.Harvester.Domain.Signals;
+using Fip.Strive.Queue.Application.Components;
+using Fip.Strive.Queue.Domain.Models;
 
 namespace Fip.Strive.Harvester.Application.Features.Classify.Services;
 
-public class ClassifyExpandWorker(IScanner scanner) : QueueWorker(SignalTypes.FileSignal)
+public class ClassifyExpandWorker(IScanner scanner) : QueueWorker((int)SignalTypes.FileSignal)
 {
     public override async Task ProcessAsync(JobDetails job, CancellationToken ct)
     {
