+++ src/harvest/Fip.Strive.Harvester.Application/Features/Expand/Services/ExpandWorker.cs
@@ -1,12 +1,11 @@
-using Fip.Strive.Core.Domain.Schemas.Queue.Enums;
-using Fip.Strive.Core.Domain.Schemas.Queue.Models;
-using Fip.Strive.Core.Domain.Schemas.Queue.Models.Signals;
-using Fip.Strive.Harvester.Application.Core.Queue.Components;
 using Fip.Strive.Harvester.Application.Features.Expand.Services.Contracts;
+using Fip.Strive.Harvester.Domain.Signals;
+using Fip.Strive.Queue.Application.Components;
+using Fip.Strive.Queue.Domain.Models;
 
 namespace Fip.Strive.Harvester.Application.Features.Expand.Services;
 
-public class ExpandWorker(IExpansionService expander) : QueueWorker(SignalTypes.ImportSignal)
+public class ExpandWorker(IExpansionService expander) : QueueWorker((int)SignalTypes.ImportSignal)
 {
     public override async Task ProcessAsync(JobDetails job, CancellationToken ct)
     {
